+++ commands/plugin.py
@@ -5,10 +5,10 @@ import sublime_plugin
 st_version = int(sublime.version())
 
 if st_version > 3000:
-    from JoomlaPack.lib.helpers import *
+    from JoomlaPack.lib import *
     from JoomlaPack.lib.extensions import Plugin
 else:
-    from lib.helpers import *
+    from lib import *
     from lib.extensions import Plugin
 
 
@@ -25,9 +25,9 @@ class NewPluginCommand(sublime_plugin.WindowCommand):
         self.show_input_panel()
 
     def show_input_panel(self):
-        show_input_panel(self.options[self.keys[self.counter]][0],
-                         self.options[self.keys[self.counter]][1],
-                         self.on_done, None, on_cancel)
+        Helper().show_input_panel(self.options[self.keys[self.counter]][0],
+                                  self.options[self.keys[self.counter]][1],
+                                  self.on_done, None, Helper().on_cancel)
 
     def on_done(self, content):
         self.options[self.keys[self.counter]][1] = content
@@ -42,28 +42,29 @@ class NewPluginCommand(sublime_plugin.WindowCommand):
 class AddFormToPluginCommand(sublime_plugin.WindowCommand):
 
     def run(self):
-        if not directories():
-            message = ('[Info] Plugin folder not found! ' +
-                       'Please, create a Joomla Plugin first.')
-            show_message('info', message)
-        else:
-            show_input_panel("Type Form name: ", "content",
-                             self.on_done, None, on_cancel)
+        Helper().show_input_panel("Type Form name: ", "content",
+                                  self.on_done, None, Helper().on_cancel)
 
     def on_done(self, name):
-        self.project = Plugin(name)
-        self.project.add_form(name)
+        self.project = Plugin()
+        # self.project.add_form(name)
+
+    def is_enabled(self):
+        return Project().has_directories()
 
 
 class AddFieldToPluginCommand(sublime_plugin.WindowCommand):
 
     def run(self):
         self.project = Plugin()
-        if directories():
-            show_input_panel("Type Field name: ", "title",
-                             self.project.on_done, None,
-                             on_cancel)
+        if Project().directories():
+            Helper().show_input_panel("Type Field name: ", "title",
+                                      self.project.on_done, None,
+                                      Helper().on_cancel)
         else:
             message = '''[Error] Project folder not found! Please,
             create a Joomla plugin first.'''
-            show_message('error', message)
+            Helper().show_message('error', message)
+
+    def is_enabled(self):
+        return Project().has_directories()