+++ lib/__init__.py
@@ -1,37 +1,15 @@
 # coding: utf-8
-
 import sublime
 
 st_version = int(sublime.version())
 if st_version > 3000:
-    from JoomlaPack.lib.helpers import window
-    from JoomlaPack.lib.helpers import settings
-    from JoomlaPack.lib.helpers import language
-    from JoomlaPack.lib.helpers import show_input_panel
-    from JoomlaPack.lib.helpers import show_message
-    from JoomlaPack.lib.helpers import on_cancel
-    from JoomlaPack.lib.helpers import directories
-    from JoomlaPack.lib.helpers import refresh
-    from JoomlaPack.lib.helpers import project_file
+    from JoomlaPack.lib.helper import Helper
+    from JoomlaPack.lib.project import Project
 else:
-    from lib.helpers import window
-    from lib.helpers import settings
-    from lib.helpers import language
-    from lib.helpers import show_input_panel
-    from lib.helpers import show_message
-    from lib.helpers import on_cancel
-    from lib.helpers import directories
-    from lib.helpers import refresh
-    from lib.helpers import project_file
+    from lib.helper import Helper
+    from lib.project import Project
 
 __all__ = [
-    'window',
-    'settings',
-    'language',
-    'show_input_panel',
-    'show_message',
-    'on_cancel',
-    'directories',
-    'refresh',
-    'project_file'
+    'Helper',
+    'Project'
 ]